+++ src/migrations/Install.php
@@ -244,6 +244,25 @@ class Install extends Migration
             );
         }
 
+        $tableSchema = Craft::$app->db->schema->getTableSchema(Constants::TABLE_NAVIGATION_DRAFT);
+        if ($tableSchema === null) {
+            $tablesCreated = true;
+            $this->createTable(
+                Constants::TABLE_NAVIGATION_DRAFT,
+                [
+                    'id'            => $this->primaryKey(),
+                    'name'          => $this->string()->notNull(),
+                    'title'         => $this->string()->notNull(),
+                    'navId'         => $this->integer()->notNull(),
+                    'site'          => $this->integer()->notNull(),
+                    'data'          => $this->mediumText()->notNull(),
+                    'dateCreated'   => $this->dateTime()->notNull(),
+                    'dateUpdated'   => $this->dateTime()->notNull(),
+                    'uid'           => $this->uid()
+                ]
+            );
+        }
+
         $tableSchema = Craft::$app->db->schema->getTableSchema(Constants::TABLE_ACTIVITY_LOG);
         if ($tableSchema === null) {
             $tablesCreated = true;
@@ -252,7 +271,7 @@ class Install extends Migration
                 [
                     'id'            => $this->primaryKey(),
                     'message'       => $this->text(),
-                    'targetclass'   => $this->string()->notNull(),
+                    'targetClass'   => $this->string()->notNull(),
                     'targetId'      => $this->integer()->notNull(),
                     'created'       => $this->dateTime()->notNull(),
                     'actions'       => $this->text()