+++ src/services/translator/GoogleTranslationService.php
@@ -42,7 +42,7 @@ class GoogleTranslationService implements TranslationServiceInterface
      */
     public function updateOrder(Order $order)
     {
-        return (new Export_ImportTranslationService([]))->updateOrder($order);
+        return (new Export_ImportTranslationService())->updateOrder($order);
     }
 
     /**
@@ -131,6 +131,6 @@ class GoogleTranslationService implements TranslationServiceInterface
 
     public function updateIOFile(Order $order, FileModel $file)
     {
-        return (new Export_ImportTranslationService([]))->updateIOFile($order, $file);
+        return (new Export_ImportTranslationService())->updateIOFile($order, $file);
     }
 }